Alarm works and exits correctly
The alarm application starts, plays music, and quits when you press
space or enter which is exactly what it is supposed to do. One note: the
alarm only rings when it has reached the second that it was when you
started it. But I do not forsee this becoming a problem.